Split the per-study arm emitters out of voting_iterations
Stage 3 of splitting voting_iterations.py: one module per study family, plus
the row-shaping helpers they all share.

  vtscore/eval/row_metrics.py       operating_metrics and what it leans on
  vtscore/eval/arms_fit_quality.py  the #3329 goodness-of-fit frame
  vtscore/eval/arms_safe_gmm.py     the #2799/#2836 cut variants
  vtscore/eval/arms_schedule.py     the #2841 mix-in schedules
  vtscore/eval/arms_fold_count.py   the #2897/#3115 fold-count arms
  vtscore/eval/arms_anchored.py     the #2852 anchored mixtures
  vtscore/eval/arms_inclusion.py    the two inclusion-budget sweeps

The four row-shaping helpers now cross a module boundary, so they lose their
underscores: _r -> round6, _folds_used -> folds_used, _honest_oracle ->
honest_oracle, _operating_metrics -> operating_metrics.  The arm emitters stay
private to the eval tier; only parse_fold_count_schedule was already public.

_cut_inclusion_arms is pinned by check-eval-app-sync.py, so its Mirror's
harness= path moves to arms_inclusion.py.  As in stage 2 the gate then passes
without --update, which is the check that the move changed no logic.

The seven inline call sites in simulate_voting_iterations stay as they are.
The issue proposed replacing them with an (enabled, emitter) dispatch; they
have genuinely different gates, argument lists, return shapes and destinations
- three feed metric_rows, the rest feed four different side sinks - so a
uniform dispatch would need a context object passed to every emitter, hiding
which inputs each arm actually reads and displacing the comments that explain
why each gate is what it is.  Only three of the seven share a shape.

voting_iterations.py is down from 5145 lines to 2388.

Refs #3405
