+++ vtscore/eval/voting_iterations.py
@@ -32,10 +32,9 @@ one.
 
 from __future__ import annotations
 
-import re
 import time
 from collections.abc import Callable
-from dataclasses import dataclass, replace
+from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Optional
 
@@ -48,8 +47,19 @@ if TYPE_CHECKING:
 from vtscore.embedding.media_vectors import media_embedding
 from vtscore.eval.al_strategies import ALContext, select_next
 from vtscore.eval.autopilot_flow import SMART_WINDOW, AutopilotFlow, app_has_detector
-from vtscore.eval.fit_quality import fit_quality_row
 from vtscore.eval.startup_schedule import StartupState, parse_startup_schedule, round_cut
+from vtscore.eval.arms_anchored import (
+    _ANCHORED_FOLD_COMBINES,
+    _ANCHORED_RULES,
+    _ANCHORED_WEIGHTS,
+    _anchored_variant_rows,
+)
+from vtscore.eval.arms_fit_quality import _fit_quality_rows
+from vtscore.eval.arms_fold_count import _fold_count_variant_rows, parse_fold_count_schedule
+from vtscore.eval.arms_inclusion import _cut_inclusion_rows, _inclusion_sweep_rows
+from vtscore.eval.arms_safe_gmm import _safe_gmm_variant_rows
+from vtscore.eval.arms_schedule import _schedule_variant_rows
+from vtscore.eval.row_metrics import operating_metrics, round6
 from vtscore.eval.step_model import (
     HEADS,
     PRODUCTION_HEAD,
@@ -67,23 +77,16 @@ from vtscore.eval.labels import evaluable_pool, media_is_positive
 from vtscore.eval.score_dumps import maybe_dump_predictions
 from vtscore.eval.voting_columns import (
     FIT_QUALITY_STRIDE_DEFAULT,
-    SKYLINE_COLUMNS,
     VOTING_COLUMNS,
 )
 from vtscore.training.blend_schedules import BlendContext
 from vtscore.training.thresholds import (
     ACQUISITION_INCLUSION_OFFSET,
-    CUT_KIND_INTERIOR,
     apply_vote_exclusion,
     FOLD_ANCHOR_QTILT_STEP,
-    FOLD_ANCHOR_WEIGHT,
     NO_GOOD_THRESHOLD,
     acquisition_inclusion,
-    anchored_gmm_fit,
     calculate_safe_threshold,
-    fold_anchored_gmm_threshold,
-    gmm_cut_from_fit,
-    rank_transfer,
     threshold_from_fold_orderings,
 )
 
@@ -234,7 +237,7 @@ def _sorted_percentile(descending: list[float], value: float) -> float:
     if not descending:
         return float("nan")
     idx = int(np.searchsorted(-np.asarray(descending, dtype=np.float64), -value, side="left"))
-    return _r(idx / len(descending))
+    return round6(idx / len(descending))
 
 
 def _blend_xcal_input(threshold: float, details: dict[str, Any]) -> float:
@@ -512,1303 +515,10 @@ def _evaluate_on_test(
         "cost": round(cost, 6),
         "fpr": round(fpr, 6),
         "fnr": round(fnr, 6),
-        **{k: _r(v) for k, v in det.items()},
+        **{k: round6(v) for k, v in det.items()},
         "auroc": round(_auroc(scores_arr, labels_arr), 6),
         "average_precision": round(_average_precision(scores_arr, labels_arr), 6),
     }
-
-
-#: ``fold_anchored[2/4]`` / ``fold_conformal_qmean[3/4]`` - *a* of *k*.
-_PROVENANCE_USED_RE = re.compile(r"\[(\d+)/(\d+)\]$")
-
-
-def _folds_used(provenance: str, k: int) -> float:
-    """How many of the *k* folds contributed a cut, read off *provenance*.
-
-    NaN for the arms where the question has no answer: the pooled conformal cut
-    and the blend take one quantile over every fold's scores at once, so no fold
-    ever "contributes a cut" that could be counted or dropped.  Reporting *k*
-    there would look like agreement with the combining arms and hide exactly the
-    asymmetry #3115 is about - a single-class fold is silently *in* the pool and
-    explicitly *out* of a mean.
-    """
-    m = _PROVENANCE_USED_RE.search(provenance)
-    return float(m.group(1)) if m else float("nan")
-
-
-def _r(x: float) -> float:
-    """Round to 6 dp when finite, else pass NaN/inf through unchanged."""
-    import math  # noqa: PLC0415
-
-    return round(x, 6) if math.isfinite(x) else x
-
-
-#: Memo for :func:`_honest_oracle`, keyed on a digest of its exact inputs.
-#: Bounded because the only reuse that matters is *within* a step, where every
-#: variant row measures the same ``(base_scores, base_labels)`` at the same
-#: inclusion: a step emits dozens of rows and the cross-fitted oracle is five
-#: sorts, so recomputing it per row would be the dominant cost of the
-#: decomposition.  Across steps the key changes and the old entry is dead, so a
-#: handful of slots is all this ever needs.
-_HONEST_ORACLE_MEMO: "dict[bytes, tuple[float, float]]" = {}
-_HONEST_ORACLE_MEMO_MAX = 8
-
-
-def _honest_oracle(scores: "np.ndarray", labels: "np.ndarray", wf: float, wn: float) -> tuple[float, float]:
-    """Memoized :func:`~vtscore.eval.transfer_rules.honest_test_oracle`.
-
-    Safe to memoize because the estimator is a pure function of its arguments -
-    its own resampling is seeded from a digest of the score array rather than
-    from global RNG state (see :func:`vtscore.eval.transfer_rules._rng`), so two
-    calls on equal inputs return bit-identical results with or without the cache.
-    The key is a digest of both arrays *and* the cost weights, because the same
-    test scores are decomposed at more than one inclusion in a single run.
-    """
-    import hashlib  # noqa: PLC0415
-
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.transfer_rules import honest_test_oracle  # noqa: PLC0415
-
-    s = np.ascontiguousarray(scores, dtype=np.float64)
-    lb = np.ascontiguousarray(labels, dtype=np.float64)
-    h = hashlib.blake2b(s.tobytes(), digest_size=16)
-    h.update(lb.tobytes())
-    h.update(np.asarray([wf, wn], dtype=np.float64).tobytes())
-    key = h.digest()
-    hit = _HONEST_ORACLE_MEMO.get(key)
-    if hit is not None:
-        return hit
-    out = honest_test_oracle(s, lb, wf, wn)
-    if len(_HONEST_ORACLE_MEMO) >= _HONEST_ORACLE_MEMO_MAX:
-        _HONEST_ORACLE_MEMO.clear()
-    _HONEST_ORACLE_MEMO[key] = out
-    return out
-
-
-def _operating_metrics(
-    scores: "np.ndarray",
-    labels: "np.ndarray",
-    threshold: float,
-    inclusion: int,
-    cal_scores: "np.ndarray | None",
-    cal_labels: "np.ndarray | None",
-    *,
-    pool_variant: str,
-    provenance: str,
-    n_pool_rows: float,
-) -> dict[str, Any]:
-    """Full per-step calibration metrics for one pooling (issue #2781).
-
-    ``scores``/``labels`` are the held-out test scores+labels under *pool_variant*
-    at the trained *threshold*.  Computes the trained cost, the oracle cost (best
-    cut on the test scores) and the resulting **regret**, plus the
-    calibration-set oracle that splits regret into *rule inefficiency*
-    (trained-vs-best-use-of-calibration) and *calibration→test shift* (best cut
-    on calibration vs. best cut on test).  ``cal_scores``/``cal_labels`` are the
-    pooled calibration fold orderings under the same pooling; ``None`` skips the
-    decomposition (leaves those columns NaN).
-
-    **Two reference points, because the naive one is optimistic** (#3116, #3248).
-    ``oracle_cost`` is the minimum of the empirical cost over the very test
-    sample it is then scored on - :func:`~vtscore.eval.calibration_metrics.oracle_cut`'s
-    own docstring calls it a lower bound on achievable cost rather than a rule -
-    so every gap measured against it is inflated by however much that minimum
-    overfits.  #2883 measured that optimism directly and found it was the *whole*
-    of the sibling ``transfer`` term (+0.041 naive, −0.001 cross-fitted).
-    ``oracle_cost_honest`` is the same quantity cross-fitted
-    (:func:`~vtscore.eval.transfer_rules.honest_test_oracle`: cut chosen on K−1
-    folds, paid on the held-out one), and the pair **brackets** the population
-    optimum rather than pinning it - naive from below, honest from above, since
-    the honest cut sees only ``(K−1)/K`` of the sample.  ``regret`` and
-    ``calibration_shift`` therefore ship beside ``regret_honest`` and
-    ``calibration_shift_honest``; read the two as an interval, and prefer the
-    honest one whenever a *level* rather than a paired contrast is being quoted.
-    ``rule_inefficiency`` is untouched by the choice - it never references the
-    test oracle - so both decompositions telescope exactly:
-
-    * ``rule_inefficiency + calibration_shift        == regret``
-    * ``rule_inefficiency + calibration_shift_honest == regret_honest``
-
-    **The split's reference moves with anything that feeds ``cal_scores``**
-    (#3116).  ``c_thr`` is estimated *from the calibration set*, so a study that
-    sweeps a knob changing that set's size or content - ``calibrate_count`` is
-    the case on record - is moving the yardstick it measures against.  As the
-    calibration set grows, ``c_thr`` converges on the test-oracle cut, which
-    shrinks ``calibration_shift`` and widens ``rule_inefficiency`` **from one
-    cause, in opposite directions**, with their sum pinned to ``regret`` by
-    construction.  #2897 read exactly that anti-correlation as a finding; it is
-    algebra.  Do not report the two terms as independent effects of such a knob
-    without a reference held fixed across the arms - and note that
-    ``rule_inefficiency`` is a signed cost gap between two cuts, **not** a
-    variance: it is routinely negative (the trained cut beating a
-    calibration-set "oracle" that overfits a handful of scores), and a study
-    asking whether the *threshold* got less variable wants ``sd(threshold)``
-    across seeds, which ``analyze_folds_2897.py`` reports.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.calibration_metrics import (  # noqa: PLC0415
-        detection_metrics,
-        inclusion_weights,
-        is_degenerate,
-        operating_cost,
-        oracle_cut,
-        threshold_percentile,
-    )
-    from vtscore.eval.label_curve import _auroc, _average_precision  # noqa: PLC0415
-
-    wf, wn = inclusion_weights(inclusion)
-    scores = np.asarray(scores, dtype=np.float64)
-    labels = np.asarray(labels, dtype=np.float64)
-
-    cost, fpr, fnr = operating_cost(scores, labels, threshold, wf, wn)
-    o_thr, o_cost, o_fpr, o_fnr = oracle_cut(scores, labels, wf, wn)
-    regret = cost - o_cost
-    # The honest half of the bracket; NaN on a test sample too small or too
-    # one-sided to cross-fit, which leaves the honest columns empty rather than
-    # silently falling back to the optimistic reference they exist to correct.
-    o_cost_honest, o_thr_honest = _honest_oracle(scores, labels, wf, wn)
-    regret_honest = cost - o_cost_honest
-
-    nan = float("nan")
-    if cal_scores is not None and np.asarray(cal_scores).size > 0:
-        cal_scores = np.asarray(cal_scores, dtype=np.float64)
-        cal_labels = np.asarray(cal_labels, dtype=np.float64)
-        c_thr, _, _, _ = oracle_cut(cal_scores, cal_labels, wf, wn)
-        cal_oracle_cost, _, _ = operating_cost(scores, labels, c_thr, wf, wn)
-        rule_inefficiency = cost - cal_oracle_cost
-        calibration_shift = cal_oracle_cost - o_cost
-        calibration_shift_honest = cal_oracle_cost - o_cost_honest
-    else:
-        c_thr = nan
-        cal_oracle_cost = nan
-        rule_inefficiency = nan
-        calibration_shift = nan
-        calibration_shift_honest = nan
-
-    return {
-        "pool_variant": pool_variant,
-        # Safe-threshold study columns (issue #2799): defaults here; the base
-        # row and the per-variant rows overwrite them where they apply.
-        "gmm_variant": "",
-        "schedule": "",
-        "xcal_threshold": _r(float(threshold)),
-        "gmm_cut": nan,
-        "blend_weight": nan,
-        # Fold-count study columns (issue #2897); only the fold-count arms set
-        # them.  ``n_cal_scores`` is the pooled calibration-set size the
-        # conformal quantile is taken over, which is what K actually buys.
-        "fold_count": nan,
-        "fold_seconds": nan,
-        # #3314: `fold_seconds` counts the fold *fits* and the conformal rule's
-        # overhead only.  A live run at K also scores the sim set once per fold
-        # and fits production's anchored mixture over the prefix, both of which
-        # scale with K and neither of which lands in `train_seconds`,
-        # `xcal_seconds`, `pool_score_seconds` or `test_score_seconds`.
-        # `cal_seconds` is the sum of all four and is what a cost ceiling reads.
-        "fold_fit_seconds": nan,
-        "fold_score_seconds": nan,
-        "anchored_seconds": nan,
-        "cal_seconds": nan,
-        "n_cal_scores": nan,
-        # #3115: how many of the K folds actually contributed a cut, parsed off
-        # the arm's own provenance.  A combine rule and a pooled quantile weight
-        # a degenerate (single-class) fold completely differently, so a contrast
-        # between them is only readable next to the count of folds that were
-        # dropped rather than averaged.
-        "n_folds_used": nan,
-        # Cut-rule study columns (issue #2836); only the variant rows set them.
-        # ``cut_fallback_kind`` says *what was substituted* where ``cut_fallback``
-        # only says *that* something was, which the two emitting families answer
-        # differently on the same fits (issue #2900).
-        "cut_fallback": 0,
-        "cut_fallback_kind": CUT_KIND_INTERIOR,
-        "cut_fail_reason": "",
-        "raw_cut_cost": nan,
-        "raw_cut_fpr": nan,
-        "raw_cut_fnr": nan,
-        "threshold": _r(float(threshold)),
-        "threshold_provenance": provenance,
-        "degenerate": 1 if is_degenerate(scores, threshold) else 0,
-        "threshold_percentile": _r(threshold_percentile(scores, threshold)),
-        "cost": _r(cost),
-        "fpr": _r(fpr),
-        "fnr": _r(fnr),
-        **{k: _r(v) for k, v in detection_metrics(scores, labels, threshold).items()},
-        "auroc": _r(float(_auroc(scores, labels))),
-        "average_precision": _r(float(_average_precision(scores, labels))),
-        "oracle_threshold": _r(float(o_thr)),
-        "oracle_cost": _r(o_cost),
-        "oracle_fpr": _r(o_fpr),
-        "oracle_fnr": _r(o_fnr),
-        "regret": _r(regret),
-        # The cross-fitted reference and the two terms it re-bases (#3116).
-        # Bracket, not replacement: `oracle_cost` bounds the population optimum
-        # from below and `oracle_cost_honest` from above, so a level quoted from
-        # either alone is one end of an interval.
-        "oracle_threshold_honest": _r(float(o_thr_honest)),
-        "oracle_cost_honest": _r(o_cost_honest),
-        "regret_honest": _r(regret_honest),
-        "cal_oracle_threshold": _r(float(c_thr)),
-        "cal_oracle_cost": _r(cal_oracle_cost),
-        "rule_inefficiency": _r(rule_inefficiency),
-        "calibration_shift": _r(calibration_shift),
-        "calibration_shift_honest": _r(calibration_shift_honest),
-        # The #3322 decomposition (see `SKYLINE_COLUMNS`).  Filled in one pass
-        # after the run, from the skyline arm's own row - a skyline is
-        # vote-independent, so there is nothing per-step to compute here, and
-        # NaN is the honest value on a run that asked for no skyline.
-        **dict.fromkeys(SKYLINE_COLUMNS, nan),
-        "n_pool_rows": _r(float(n_pool_rows)),
-    }
-
-
-#: Safe-threshold cut variants (issues #2799, #2836): ``(name, fit_scores, rule)``.
-#: ``fit_scores`` picks which sim-set score distribution the mixture is fitted on
-#: ("pooled" = the style's inference max-pool, what production fits post-#2797;
-#: "image" = the whole-image vector scores, the historical pre-#2797 geometry).
-#: ``rule`` names a cut in :mod:`vtscore.eval.cut_rules` - the ``lam``-tilt
-#: family over the Gaussian mixture ("mid" is production; "cross" is #2798's
-#: count-optimal crossing, reverted by #2833; "priorfree"/"rate" are #2836's
-#: rate-optimal tilts), the same tilts over a Gumbel-low mixture ("gumbel_*") and
-#: over one whose Gumbel may land on either mode ("gumbel_any_*", #2846's repair
-#: to the first family's fallback rate), and the two label-reading diagnostics
-#: ("supervised", "sim_oracle") that locate the error rather than compete to
-#: ship.  ``xcal_only`` is the no-blend control: the conformal threshold at the
-#: same step, or - on a step whose folds fell back - the same sentinel the
-#: shipped blend feeds its x-cal side (:func:`_blend_xcal_input`), so the
-#: control is the blend's own input with the mix-in removed rather than a cut
-#: nobody computed.  ``pooled_mid`` must reproduce the production blend exactly.
-#: The ``tail_a*`` sweep is #2881's one-constant rule at seven tail levels - the
-#: fitted Bad component's own quantile rather than a crossing of any kind.
-#:
-#: **The rule names here are the ones in :mod:`vtscore.eval.cut_rules`**, spelled
-#: out rather than derived, because this module is deliberately import-light (no
-#: numpy at import time) and importing the rule tables to build the list would
-#: undo that.  ``test_cut_rules`` asserts the two agree, so the duplication
-#: cannot drift silently - which matters more than usual here, since a rule that
-#: is defined but never emitted produces a table with a missing row rather than
-#: an error.
-#:
-#: The #2798 logit-space variants are gone: #2799 measured them at +0.0006 cost
-#: (dead) and each extra fit costs a step's CPU that the #2836 arms need.
-_SAFE_GMM_VARIANTS: tuple[tuple[str, str, str], ...] = (
-    ("xcal_only", "", ""),
-    ("image_mid", "image", "mid"),
-    ("image_cross", "image", "cross"),
-    ("image_priorfree", "image", "priorfree"),
-    ("image_rate", "image", "rate"),
-    ("pooled_mid", "pooled", "mid"),
-    ("pooled_cross", "pooled", "cross"),
-    ("pooled_priorfree", "pooled", "priorfree"),
-    ("pooled_rate", "pooled", "rate"),
-    ("pooled_gumbel_cross", "pooled", "gumbel_cross"),
-    ("pooled_gumbel_priorfree", "pooled", "gumbel_priorfree"),
-    ("pooled_gumbel_rate", "pooled", "gumbel_rate"),
-    ("pooled_gumbel_any_cross", "pooled", "gumbel_any_cross"),
-    ("pooled_gumbel_any_priorfree", "pooled", "gumbel_any_priorfree"),
-    ("pooled_gumbel_any_rate", "pooled", "gumbel_any_rate"),
-    ("pooled_tail_a040", "pooled", "tail_a040"),
-    ("pooled_tail_a080", "pooled", "tail_a080"),
-    ("pooled_tail_a110", "pooled", "tail_a110"),
-    ("pooled_tail_a158", "pooled", "tail_a158"),
-    ("pooled_tail_a220", "pooled", "tail_a220"),
-    ("pooled_tail_a300", "pooled", "tail_a300"),
-    ("pooled_tail_a400", "pooled", "tail_a400"),
-    ("pooled_supervised", "pooled", "supervised"),
-    ("pooled_sim_oracle", "pooled", "sim_oracle"),
-    # #2883: the last link's shape.  Four subsample levels give the learning
-    # curve in sim-set size (the test set and the trajectory are identical
-    # across them - only the number of labelled sim scores moves), and two
-    # variance-reduced estimators of the same target test whether the empirical
-    # minimiser is the bound `family_headroom_exhausted` treats it as.
-    ("pooled_sim_oracle_f050", "pooled", "sim_oracle_f050"),
-    ("pooled_sim_oracle_f100", "pooled", "sim_oracle_f100"),
-    ("pooled_sim_oracle_f250", "pooled", "sim_oracle_f250"),
-    ("pooled_sim_oracle_f500", "pooled", "sim_oracle_f500"),
-    ("pooled_sim_oracle_bag", "pooled", "sim_oracle_bag"),
-    ("pooled_sim_oracle_smooth", "pooled", "sim_oracle_smooth"),
-    # The label-free counterpart: bag the mixture fit rather than the labelled
-    # cost curve.  Exploratory, not ship-gated - see PREREG.
-    ("pooled_bagfit_mid", "pooled", "bagfit_mid"),
-    ("pooled_bagfit_priorfree", "pooled", "bagfit_priorfree"),
-)
-
-#: Variants that read the sim set's true labels.  Reported for the decomposition,
-#: never eligible to ship - a rule cannot see these labels in the app.
-_ORACLE_VARIANTS: frozenset[str] = frozenset(
-    {
-        "pooled_supervised",
-        "pooled_sim_oracle",
-        # #2883's readings of the same sim set - label-reading for the same
-        # reason and, like the two above, emitting NaN rather than falling back
-        # to a midpoint under another rule's name.
-        "pooled_sim_oracle_f050",
-        "pooled_sim_oracle_f100",
-        "pooled_sim_oracle_f250",
-        "pooled_sim_oracle_f500",
-        "pooled_sim_oracle_bag",
-        "pooled_sim_oracle_smooth",
-    }
-)
-
-
-def _fit_quality_rows(
-    base_row: dict[str, Any],
-    safe_cut: Any,
-    sim_scores_by_geometry: dict[str, Any],
-    sim_labels_by_geometry: dict[str, Any],
-    threshold: float,
-) -> list[dict[str, Any]]:
-    """Goodness-of-fit rows for one step (issue #3329).
-
-    Two scopes, deliberately not pooled - see :data:`FIT_QUALITY_ROW_COLUMNS`.
-
-    The **fold** scopes read ``safe_cut.fits`` against ``safe_cut.fold_haystacks``,
-    which is the shipped estimator scored against its own data.  Nothing else in
-    the eval tier reads those fits at all, so a fold row is the only place the
-    threshold the app actually computes is compared to the distribution it claims
-    to describe.  They are label-free: a fold haystack is the unlabelled
-    remainder under *that fold's* model, and the sim labels belong to the final
-    model's score scale, so attaching them here would compare two scalings.
-
-    The **sim** scopes fit the unanchored mixture to the labelled sim scores and
-    carry the shape and identification statistics.  Both pooling geometries are
-    emitted because ``pooled`` and ``image`` are the same media under the same
-    model with only the pooling changed - the exactly-paired form of "does
-    max-pooling make the Bad mode non-Gaussian?".
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.training.thresholds import fit_score_gmm, gmm_fit_array  # noqa: PLC0415
-
-    rows: list[dict[str, Any]] = []
-
-    if safe_cut is not None:
-        fits = getattr(safe_cut, "fits", ()) or ()
-        haystacks = getattr(safe_cut, "fold_haystacks", ()) or ()
-        anchor_counts = getattr(safe_cut, "anchor_counts", ()) or ()
-        for i, (fit, hay) in enumerate(zip(fits, haystacks, strict=False)):
-            arr = np.asarray(hay, dtype=np.float64).ravel()
-            # The votes THIS fold anchored on - its held-out share - which is
-            # what the mass fraction needs.  The obvious-looking `n_anchored` is
-            # a count of FOLDS (0/1/2), and passing it produced a mass share
-            # flat at 2.9e-4 across every click of the first real run, off by
-            # the fold's vote count (#3329).  Recorded on the cut rather than
-            # recomputed here so the two cannot disagree.
-            n_anchors = int(anchor_counts[i]) if i < len(anchor_counts) else 0
-            # The counterfactual the H3 delta is against: the same sample, the
-            # same estimator, no anchors.  `fold_haystacks[i]` IS the array the
-            # shipped fit was fitted to (sorted), so the two fits are comparable
-            # point for point.
-            unanchored = fit_score_gmm(arr)
-            fq = fit_quality_row(
-                arr,
-                fit,
-                cut=threshold,
-                n_anchors=n_anchors,
-                anchor_weight=float(getattr(safe_cut, "anchor_weight", FOLD_ANCHOR_WEIGHT)),
-                unanchored_fit=unanchored,
-            )
-            rows.append(
-                {
-                    **base_row,
-                    "scope": f"fold{i}",
-                    "fold_index": i,
-                    "n_folds": len(fits),
-                    "n_fit_sample": int(arr.size),
-                    "fit_ok": fit is not None,
-                    "fq_w_lo": fit.w_lo,
-                    "fq_mu_lo": fit.mu_lo,
-                    "fq_var_lo": fit.var_lo,
-                    "fq_w_hi": fit.w_hi,
-                    "fq_mu_hi": fit.mu_hi,
-                    "fq_var_hi": fit.var_hi,
-                    "fq_cut": float(threshold),
-                    **fq,
-                }
-            )
-
-    for geometry, scores in sim_scores_by_geometry.items():
-        if scores is None:
-            continue
-        labels = sim_labels_by_geometry.get(geometry)
-        arr = gmm_fit_array(np.asarray(scores, dtype=np.float64).ravel())
-        fit = fit_score_gmm(arr)
-        fq = fit_quality_row(
-            arr,
-            fit,
-            cut=threshold,
-            labels=labels,
-            label_scores=np.asarray(scores, dtype=np.float64).ravel() if labels is not None else None,
-        )
-        rows.append(
-            {
-                **base_row,
-                "scope": f"sim:{geometry}",
-                "fold_index": -1,
-                "n_folds": 0,
-                "n_fit_sample": int(arr.size),
-                "fit_ok": fit is not None,
-                "fq_w_lo": None if fit is None else fit.w_lo,
-                "fq_mu_lo": None if fit is None else fit.mu_lo,
-                "fq_var_lo": None if fit is None else fit.var_lo,
-                "fq_w_hi": None if fit is None else fit.w_hi,
-                "fq_mu_hi": None if fit is None else fit.mu_hi,
-                "fq_var_hi": None if fit is None else fit.var_hi,
-                "fq_cut": float(threshold),
-                **fq,
-            }
-        )
-    return rows
-
-
-def _safe_gmm_variant_rows(
-    details: dict[str, Any],
-    base_scores: "np.ndarray",
-    base_labels: "np.ndarray",
-    sim_scores_by_geometry: dict[str, list[float]],
-    sim_labels_by_geometry: dict[str, "np.ndarray"],
-    inclusion: int,
-    n_pool_rows: float,
-    schedule: str | None = None,
-) -> tuple[list[dict[str, Any]], list[dict[str, Any]]]:
-    """Metric rows per cut variant, plus the per-geometry decomposition rows.
-
-    Every variant re-cuts the *same* per-step model: each candidate sim-set score
-    distribution is fitted once (both mixture families), every cut rule reads that
-    one fit, and each cut is blended with the step's conformal threshold on the
-    production label ramp.  All variants are evaluated against the same held-out
-    test scores (*base_scores*, the inference max-pool), so the rows are paired
-    within a step by construction.
-
-    Two costs are recorded per variant: ``cost`` at the *blended* threshold (what
-    a user would get, and what the ship decision reads) and ``raw_cut_cost`` at
-    the unblended cut (what the *rule* is worth, undamped by the conformal
-    threshold it is averaged with).  On the ramp the blend can shrink a large cut
-    difference to a small cost difference, so the rule comparison belongs on the
-    raw column and the ship comparison on the blended one.
-
-    A rule whose root does not exist on a given fit falls back to that fit's
-    midpoint and is flagged in ``cut_fallback`` so the analyzer can exclude
-    fallen-back steps from a rule's own contrast rather than silently scoring
-    the midpoint under another name.  **The midpoint is this family's fallback,
-    not production's** - the shipped ``rate`` rule
-    (:func:`~vtscore.training.thresholds.gmm_cut_from_fit`) continues past the
-    inter-mean interval at its own first-order slope instead, so on the fits
-    where this flag fires these arms are measuring a different rule than the
-    app runs.  That divergence is **kept on purpose** (issue #2900): this family
-    compares tilts against each other on one fit, and a rule-independent
-    stand-in is what keeps ``rate`` commensurable with the ``cross`` and
-    ``priorfree`` siblings it is differenced against - at inclusion 0 it is what
-    keeps ``rate`` bit-identical to ``priorfree``, which is how every report in
-    ``docs/experiments/2026-08-04-gmm-cut/`` reads those rows.  It is no longer *invisible*
-    though: ``cut_fallback_kind`` carries
-    :data:`~vtscore.eval.cut_rules.CUT_KIND_MIDPOINT` on exactly these steps,
-    against the production family's ``continued`` / ``degenerate_midpoint``, so
-    an analysis that wants the shipped path can filter for it instead of reading
-    a substituted midpoint as "what the app would have done".  The fold-anchored
-    family below calls the production function directly and so does not have
-    that gap at all.  For the EVT rules
-    ``cut_fail_reason`` additionally names *which* guard declined, because the
-    repairs those guards want are different and the counts alone cannot tell them
-    apart (issue #2846).  The oracle variants do not fall back; they emit NaN cuts
-    and are dropped by the analyzer's joins.
-
-    Returns ``(variant_rows, diagnostic_rows)``; the diagnostic rows carry the
-    fitted mixture parameters and every cut in the decomposition chain, one row
-    per (step, geometry), and still need the caller's identifying columns.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.calibration_metrics import inclusion_weights, operating_cost, oracle_cut  # noqa: PLC0415
-    from vtscore.eval.cut_rules import CUT_KIND_MIDPOINT, decomposition_cuts  # noqa: PLC0415
-    from vtscore.eval.transfer_rules import honest_test_oracle  # noqa: PLC0415
-    from vtscore.training.thresholds import blend_gmm_threshold, safe_blend_weight  # noqa: PLC0415
-
-    xcal = float(details["xcal_threshold"])
-    n_votes = int(details["n_votes"])
-    pre_blend_provenance = str(details.get("pre_blend_provenance", "conformal"))
-    fold_orderings = details.get("fold_orderings") or []
-    cal_scores = np.array([s for scores, _ in fold_orderings for s in scores]) if fold_orderings else None
-    cal_labels = np.array([lb for _, labels_ in fold_orderings for lb in labels_]) if fold_orderings else None
-    wf, wn = inclusion_weights(inclusion)
-    nan = float("nan")
-
-    # One fit pass per geometry; every rule below reads these.
-    cuts_by_geometry: dict[str, dict[str, float]] = {}
-    reasons_by_geometry: dict[str, dict[str, str]] = {}
-    diag_rows: list[dict[str, Any]] = []
-    geometries = sorted({fit for _n, fit, _r in _SAFE_GMM_VARIANTS if fit})
-    for geometry in geometries:
-        scores = sim_scores_by_geometry[geometry]
-        labels = sim_labels_by_geometry[geometry]
-        if len(scores) < 2:
-            # Mirrors calculate_gmm_threshold's "too few scores" default so the
-            # production-blend sanity check holds at every step.
-            cuts_by_geometry[geometry] = dict.fromkeys((r for _n, f, r in _SAFE_GMM_VARIANTS if f == geometry), 0.5)
-            reasons_by_geometry[geometry] = {}
-            continue
-        cuts, params, reasons = decomposition_cuts(scores, labels, wf, wn)
-        cuts_by_geometry[geometry] = cuts
-        reasons_by_geometry[geometry] = reasons
-        diag = {
-            "geometry": geometry,
-            "sim_prevalence": _r(float(np.mean(labels))) if len(labels) else nan,
-            # The count, not just the rate: a threshold estimated from labelled
-            # scores is limited by the *rarer* class, so #2883's scaling claim is
-            # about positives and prevalence alone cannot express it.
-            "sim_n_pos": _r(float(np.sum(labels == 1.0))) if len(labels) else nan,
-        }
-        # ``evt_fit_fail`` is a reason string; everything else in params is numeric.
-        diag.update({k: v if isinstance(v, str) else _r(float(v)) for k, v in params.items()})
-        diag.update({f"tau_{name}": _r(float(value)) for name, value in cuts.items()})
-        diag["tau_test_oracle"] = nan  # filled below, once the test oracle is known
-        diag_rows.append(diag)
-
-    ctx = BlendContext(
-        n_labels=n_votes,
-        n_good=int(details.get("n_good", 0)),
-        n_bad=int(details.get("n_bad", n_votes)),
-    )
-    weight = safe_blend_weight(ctx, schedule)
-    rows: list[dict[str, Any]] = []
-    for name, geometry, rule in _SAFE_GMM_VARIANTS:
-        fallback = 0
-        fallback_kind = CUT_KIND_INTERIOR
-        fail_reason = ""
-        if name == "xcal_only":
-            threshold = xcal
-            gmm_cut = nan
-            provenance = pre_blend_provenance
-        else:
-            gmm_cut = cuts_by_geometry[geometry][rule]
-            if not np.isfinite(gmm_cut) and name not in _ORACLE_VARIANTS:
-                gmm_cut = cuts_by_geometry[geometry].get("mid", nan)
-                fallback = 1
-                fallback_kind = CUT_KIND_MIDPOINT
-                # Empty for the Gaussian rules, which have no reason vocabulary;
-                # the EVT rules name the guard that declined so a fallback can be
-                # attributed rather than merely counted (issue #2846).
-                fail_reason = reasons_by_geometry[geometry].get(rule, "")
-            threshold = blend_gmm_threshold(xcal, gmm_cut, ctx, schedule=schedule) if np.isfinite(gmm_cut) else nan
-            provenance = "gmm_blend"
-        if not np.isfinite(threshold):
-            continue
-        row = _operating_metrics(
-            base_scores,
-            base_labels,
-            threshold,
-            inclusion,
-            cal_scores,
-            cal_labels,
-            pool_variant="max",
-            provenance=provenance,
-            n_pool_rows=n_pool_rows,
-        )
-        row["gmm_variant"] = name
-        row["schedule"] = ""
-        row["xcal_threshold"] = _r(xcal)
-        row["gmm_cut"] = _r(gmm_cut)
-        row["blend_weight"] = _r(weight)
-        row["cut_fallback"] = fallback
-        row["cut_fallback_kind"] = fallback_kind
-        row["cut_fail_reason"] = fail_reason
-        if np.isfinite(gmm_cut):
-            raw_cost, raw_fpr, raw_fnr = operating_cost(base_scores, base_labels, gmm_cut, wf, wn)
-            row["raw_cut_cost"] = _r(raw_cost)
-            row["raw_cut_fpr"] = _r(raw_fpr)
-            row["raw_cut_fnr"] = _r(raw_fnr)
-        rows.append(row)
-
-    # The last link in the chain: the best cut on the held-out test set.  Read off
-    # any emitted row (all share the same base_scores/base_labels oracle).
-    if rows and diag_rows:
-        # #2883: that cut is the argmin of the empirical cost on the test sample
-        # *itself*, so its cost is a sample minimum - biased low, which biases
-        # `transfer` high by however much the reference overfits.  Record the
-        # cross-fitted version beside it (cut and cost on disjoint folds) so the
-        # last link can be reported as a bracket instead of a point.
-        honest_cost, honest_tau = honest_test_oracle(base_scores, base_labels, wf, wn)
-        # From `oracle_cut` itself, NOT by re-scoring at `rows[0]["oracle_threshold"]`:
-        # that column is rounded on the way out, and re-evaluating a cost at a
-        # rounded threshold moves items across the boundary.  With ~55 test
-        # positives one FNR step is 1/55 = 0.018 - half the size of the term this
-        # study is measuring - so the rounding is not a rounding error here.
-        _naive_tau, naive_cost, _nfpr, _nfnr = oracle_cut(base_scores, base_labels, wf, wn)
-        n_test = float(np.asarray(base_labels).size)
-        n_test_pos = float(np.sum(np.asarray(base_labels) == 1.0))
-        for diag in diag_rows:
-            diag["tau_test_oracle"] = rows[0]["oracle_threshold"]
-            diag["tau_test_oracle_honest"] = _r(honest_tau)
-            diag["cost_test_oracle_naive"] = _r(naive_cost)
-            diag["cost_test_oracle_honest"] = _r(honest_cost)
-            diag["test_n"] = _r(n_test)
-            diag["test_n_pos"] = _r(n_test_pos)
-    return rows, diag_rows
-
-
-def _schedule_variant_rows(
-    details: dict[str, Any],
-    base_scores: "np.ndarray",
-    base_labels: "np.ndarray",
-    sim_pooled_scores: list[float],
-    inclusion: int,
-    n_pool_rows: float,
-    schedules: list[str],
-) -> list[dict[str, Any]]:
-    """One metric row per mix-in **schedule** (issue #2841).
-
-    The dual of :func:`_safe_gmm_variant_rows`, which holds the schedule fixed
-    and varies the GMM *cut*: here the cut is production's (midpoint of the
-    component means, fitted on the style's inference pool) and the *schedule*
-    varies.  Every schedule re-combines the same two candidate cuts from the
-    same per-step model against the same held-out test scores, so the rows are
-    paired within a step by construction and differ only in the mix-in rule.
-
-    This is the study's **screen**, not its verdict.  Holding the trajectory
-    fixed is precisely what makes it cheap - one simulation scores every
-    schedule - but the blended threshold also feeds acquisition (Autopilot's
-    Hard phase picks the item nearest the decision boundary), so schedules that
-    would have labelled *different items* cannot show that here.  The A/B runs
-    exist to measure what this screen structurally cannot see.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.training.thresholds import (  # noqa: PLC0415
-        blend_gmm_threshold,
-        fit_gmm_threshold,
-        safe_blend_weight,
-    )
-
-    xcal = float(details["xcal_threshold"])
-    n_votes = int(details["n_votes"])
-    pre_blend_provenance = str(details.get("pre_blend_provenance", "conformal"))
-    fold_orderings = details.get("fold_orderings") or []
-    cal_scores = np.array([s for scores, _ in fold_orderings for s in scores]) if fold_orderings else None
-    cal_labels = np.array([lb for _, labels_ in fold_orderings for lb in labels_]) if fold_orderings else None
-
-    # Required, not defaulted: the ``rare``/``pos`` families ramp on the class
-    # split, so a guessed split would silently mis-score two whole families
-    # rather than fail.  The caller sets both keys alongside ``n_votes``.
-    ctx = BlendContext(n_labels=n_votes, n_good=int(details["n_good"]), n_bad=int(details["n_bad"]))
-    # One fit, re-combined under every schedule.  The corridor schedules need
-    # the component means, so the fit object rides along with the cut.
-    gmm_cut, gmm_fit = fit_gmm_threshold(sim_pooled_scores)
-
-    rows: list[dict[str, Any]] = []
-    for name in schedules:
-        threshold = blend_gmm_threshold(xcal, gmm_cut, ctx, schedule=name, fit=gmm_fit)
-        weight = safe_blend_weight(ctx, name)
-        row = _operating_metrics(
-            base_scores,
-            base_labels,
-            threshold,
-            inclusion,
-            cal_scores,
-            cal_labels,
-            pool_variant="max",
-            provenance="gmm_blend" if weight < 1.0 else pre_blend_provenance,
-            n_pool_rows=n_pool_rows,
-        )
-        row["gmm_variant"] = ""
-        row["schedule"] = name
-        row["xcal_threshold"] = _r(xcal)
-        row["gmm_cut"] = _r(gmm_cut)
-        row["blend_weight"] = _r(weight)
-        rows.append(row)
-    return rows
-
-
-def _fold_schedule_segment(part: str) -> tuple[int, int]:
-    """One ``"K@N"`` segment as ``(cut, count)``, or a ValueError naming it.
-
-    Split out of :func:`parse_fold_count_schedule` so the validation reads as a
-    list of the four ways a spec can be wrong rather than as branches inside the
-    parse loop.  The error says which SEGMENT was bad, because a schedule is
-    typed into a launcher and "invalid schedule" would leave a reader diffing
-    the string by eye.
-    """
-    k_str, sep, n_str = part.strip().partition("@")
-    if not sep:
-        raise ValueError(f"fold-count schedule segment {part!r} is not 'K@N'")
-    try:
-        k, n = int(k_str), int(n_str)
-    except ValueError as exc:
-        raise ValueError(f"fold-count schedule segment {part!r} is not 'K@N': {exc}") from exc
-    if k < 1:
-        raise ValueError(f"fold-count schedule segment {part!r}: K must be >= 1")
-    if n < 1:
-        raise ValueError(f"fold-count schedule segment {part!r}: N must be >= 1")
-    return n, k
-
-
-def parse_fold_count_schedule(spec: str | None, base: int) -> "Callable[[int], int] | None":
-    """Eval-only (#3314): resolve ``calibrate_count`` per step from the vote count.
-
-    ``"K@N"`` means ``K(n_votes) = K while n_votes < N, else base`` - the family
-    pre-registered in ``docs/experiments/2026-08-28-calibration-fold-count-3310/PLAN.md``,
-    written the issue's way round: *more folds while the labelset is small,
-    decaying to production's count*.  Several segments may be chained
-    (``"8@25,4@60"``), and they are read in ascending ``N`` order, so the first
-    cut a vote count falls under wins and the ordering in the string does not
-    matter.
-
-    This is a **harness knob and not a shipped setting**.  ``CALIBRATE_COUNT``
-    and the app's own constant are untouched, so no other study's arm moves:
-    a run that does not set it resolves to *base* at every step, which is
-    exactly what the constant did before.  If the schedule ever ships, the
-    right shape is a ``production_fold_count_for(n_votes)`` beside
-    ``production_split_for``, with ``scripts/check-eval-app-sync.py`` gaining
-    the mirror - the same discipline #3287's split fraction went through.
-
-    Returns ``None`` for an empty spec, so the caller's fast path is the
-    unscheduled one.
-    """
-    if not spec or not spec.strip():
-        return None
-    segments = sorted(_fold_schedule_segment(part) for part in spec.split(",") if part.strip())
-    if not segments:
-        return None
-
-    def resolve(n_votes: int) -> int:
-        for cut, k in segments:
-            if n_votes < cut:
-                return k
-        return base
-
-    return resolve
-
-
-def _stop(timings: dict[str, float] | None, started: float) -> None:
-    """Record ``anchored_seconds`` since *started*, when the caller wants timings."""
-    if timings is not None:
-        timings["anchored_seconds"] = time.monotonic() - started
-
-
-def _fold_count_arms(
-    prefix: list[tuple[list[float], list[float]]],
-    xcal: float,
-    inclusion: int,
-    haystacks: "list[np.ndarray] | None",
-    final_fit_scores: list[float] | None,
-    ctx: BlendContext | None,
-    gmm_cut: float | None,
-    gmm_fit: Any,
-    schedule: str | None,
-    timings: dict[str, float] | None = None,
-) -> list[tuple[str, float, str, float]]:
-    """``(arm, threshold, provenance, blend weight)`` for one fold prefix.
-
-    *timings*, when given, collects the wall clock of the parts a **live** run
-    at this K would actually pay for (#3314) - currently ``anchored_seconds``,
-    the fit of production's fold-anchored mixture over this prefix.  The other
-    arms here are counterfactual re-cuts of the same folds and cost a live run
-    nothing, so they are deliberately not timed: a cost model built off the
-    whole function would price the study's own instrumentation as if the user
-    waited through it.
-
-    Split out of :func:`_fold_count_variant_rows` so the arm table is one
-    readable list rather than a branch pile inside the K loop; every arm here is
-    a different *rule* applied to the **same** already-trained folds, so adding
-    one costs arithmetic and no fits.
-
-    ``haystacks`` is ``None`` when the step does not carry one sim-set score
-    array per fold in the prefix, which gates every arm that has to read a cut
-    in the fold's own distribution (:func:`~vtscore.training.thresholds.rank_transfer`).
-    ``final_fit_scores`` is the final model's haystack the quantile-realizing
-    arms cut on - under the #3308 convention the caller passes it with the
-    voted items already dropped, matching the fold ``haystacks``; the ``blend``
-    arm's GMM inputs (*gmm_cut*, *gmm_fit*) are fitted upstream on the full
-    distribution, as production's fallback is.
-    """
-    import dataclasses  # noqa: PLC0415
-
-    from vtscore.training.thresholds import (  # noqa: PLC0415
-        FOLD_CONFORMAL_COMBINES,
-        blend_gmm_threshold,
-        combined_fold_conformal_threshold,
-        fit_fold_anchored_cut,
-        fold_anchored_gmm_threshold,
-        safe_blend_weight,
-    )
-
-    nan = float("nan")
-    arms: list[tuple[str, float, str, float]] = [("xcal", xcal, "conformal", nan)]
-    if ctx is not None and gmm_cut is not None:
-        weight = safe_blend_weight(ctx, schedule)
-        blended = blend_gmm_threshold(xcal, gmm_cut, ctx, schedule=schedule, fit=gmm_fit)
-        arms.append(("blend", blended, "gmm_blend" if weight < 1.0 else "conformal", weight))
-
-    # #3115, the combine rule.  ``xcal`` above IS the pooled control - it calls
-    # `threshold_from_fold_orderings` verbatim - so these are challengers rather
-    # than a re-emission of it.  The score-space pair needs nothing the pooled
-    # arm does not already have; the quantile-space pair needs a haystack per
-    # fold, the same condition the anchored arms carry.
-    #
-    # `transferable` binds that condition once - "can this arm read a cut in a
-    # fold's own scale?" - so it is asked in one place instead of re-derived at
-    # each use site, and so both `None` cases narrow for the type checker.
-    final_scores = final_fit_scores if final_fit_scores else None
-    transferable = haystacks if (haystacks is not None and final_scores is not None) else None
-    for combine in FOLD_CONFORMAL_COMBINES:
-        quantile_space = combine.startswith("q")
-        if quantile_space and transferable is None:
-            continue
-        value, prov = combined_fold_conformal_threshold(
-            prefix,
-            inclusion,
-            combine=combine,
-            fold_haystacks=transferable if quantile_space else None,
-            final_scores=final_scores if quantile_space else None,
-        )
-        arms.append((combine, value, prov, nan))
-
-    # The arms for the rule users actually get.  ``anchored`` is production
-    # (`FOLD_ANCHOR_COMBINE`); ``anchored_qmedian`` re-cuts the *same* fit under
-    # the robust combine, which puts #3115's contamination question on the
-    # shipped path rather than only on the retired blend.
-    #
-    # Fitted **once** and re-cut, not fitted twice.  `FoldAnchoredCut` exists to
-    # separate the fit from the cut, and the combine rule is read at cut time, so
-    # a second `fold_anchored_gmm_threshold` call would re-run one anchored EM
-    # per fold to reach the same mixtures - doubling the study's dominant cost
-    # (sum over the K grid, so 52 EM fits per step at K<=16, not 16).  It is also
-    # the stronger contrast: the two rows differ in the combine and in *nothing
-    # else*, including the fits' own numerical noise.
-    if transferable is not None and final_scores is not None:
-        # Timed with the fallback inside it (#3314): a step whose anchored fit
-        # degenerates still pays for the attempt *and* for the ladder the
-        # shipped helper walks afterwards, and that is what a user waits
-        # through.  Charging those steps only for the branch they did not take
-        # would under-price exactly the cold-start regime this study reads.
-        t_anchored = time.monotonic()
-        cut = fit_fold_anchored_cut(transferable, prefix, final_scores)
-        if cut is None:
-            # Both arms land on the same terminal fallback; take it from the
-            # shipped helper rather than duplicating its ladder here.
-            value, prov = fold_anchored_gmm_threshold(transferable, prefix, final_scores, inclusion)
-            _stop(timings, t_anchored)
-            arms.extend([("anchored", value, prov, nan), ("anchored_qmedian", value, prov, nan)])
-        else:
-            shipped = cut.threshold_at(inclusion)
-            # The clock stops HERE, on production's own cut and before the
-            # `qmedian` re-cut: that arm is a counterfactual read of the same
-            # fitted mixture and costs a live run nothing, so timing it would
-            # charge K for the harness's own instrumentation.  (Sharing the fit
-            # is exactly why the two arms are cheap; see the note above.)
-            _stop(timings, t_anchored)
-            arms.append(("anchored", shipped, cut.provenance, nan))
-            robust = dataclasses.replace(cut, combine="qmedian")
-            arms.append(("anchored_qmedian", robust.threshold_at(inclusion), robust.provenance, nan))
-    return arms
-
-
-def _fold_count_variant_rows(
-    details: dict[str, Any],
-    base_scores: "np.ndarray",
-    base_labels: "np.ndarray",
-    inclusion: int,
-    n_pool_rows: float,
-    counts: list[int],
-    sim_pooled_scores: list[float] | None,
-    schedule: str | None,
-    sim_fit_scores: list[float] | None = None,
-) -> list[dict[str, Any]]:
-    """One metric row per calibration **fold count** K (issue #2897).
-
-    *sim_fit_scores* is the sim haystack under the #3308 population convention
-    (voted items dropped); the ``anchored`` / quantile-space arms realize their
-    cuts on it, mirroring the shipped rule, while the retired ``blend`` arm
-    keeps fitting *sim_pooled_scores* - production's fallback blend also keeps
-    the full distribution.  ``None`` falls back to *sim_pooled_scores*.
-
-    The study's screen for "does more cross-calibration buy anything, and what
-    does it cost".  It is exact rather than approximate, because the folds are
-    *nested*: :func:`~vtscore.training.thresholds.compute_fold_orderings` draws
-    each fold as an independent stratified split off one ``RandomState(42)``
-    stream, at a per-fold size that does not depend on the count, so the K folds
-    a live ``calibrate_count=K`` run would train are byte-for-byte the first K
-    of the Kmax folds trained here.  Slicing the prefix therefore reproduces
-    each K's threshold exactly, and the arm at ``K == calibrate_count``
-    reproduces this step's own pre-blend conformal cut - the control that
-    licenses the rest of the table.
-
-    Arms per K, because the fold count and the shipped threshold are different
-    questions - and, since #3115, because *how* the folds are combined is a
-    third one.  Every arm below re-reads the same already-trained fold prefix,
-    so the whole table costs arithmetic:
-
-    * ``folds_k{K}_xcal`` - the raw cross-calibration cut, the thing K is
-      actually a knob on.
-    * ``folds_k{K}_blend`` - that cut after the ``cap50`` safe-threshold mix-in.
-      The blend weight depends only on the vote counts, so it is identical
-      across K and this arm isolates how much of K's benefit survives being
-      averaged with the GMM cut.  Emitted only when the step has the pooled sim
-      scores the blend fits.
-    * ``folds_k{K}_anchored`` - **production's rule** (#3116).  The blend above
-      was retired by the 2026-08-05 population-anchored run; the shipped path is
-      :func:`~vtscore.training.thresholds.fold_anchored_gmm_threshold`, which
-      fits one anchored mixture *per fold* and combines them in quantile space.
-      K therefore moves the shipped threshold through a path the other two arms
-      do not exercise at all - the blend's GMM half is a single unanchored fit
-      on the sim haystack and is K-independent by construction, so ``blend``
-      varies only in its x-cal half.  Without this arm every conclusion about
-      what ``calibrate_count`` does to the threshold users actually get is
-      partial.  Emitted when the step carries at least K fold haystacks, which
-      :func:`_safe_threshold_for_step` supplies (so: under safe thresholds).
-    * ``folds_k{K}_{tmean,tmedian,qmean,qmedian}`` - **the combine rule**
-      (#3115).  ``xcal`` above *is* the pooled control: it calls
-      :func:`~vtscore.training.thresholds.threshold_from_fold_orderings`
-      verbatim, which pools every fold's held-out scores into one bag and takes
-      a single conformal quantile.  These four take one cut per fold and combine
-      them instead, in score space (``t*``) or in fold-quantile space (``q*``);
-      see :data:`~vtscore.training.thresholds.FOLD_CONFORMAL_COMBINES` for what
-      each leg of the contrast isolates.  ``q*`` needs a haystack per fold and
-      so shares the ``anchored`` arm's condition.
-    * ``folds_k{K}_anchored_qmedian`` - production's rule re-cut under the
-      robust combine, which puts the contamination question on the **shipped**
-      path and not only on the retired blend.
-
-    Note what is *not* a defect here: the unanchored ``fit_gmm_threshold`` is
-    hoisted out of the K loop, and that is correct rather than a frozen
-    reference - it reads only ``sim_pooled_scores``, which no fold count
-    touches, so re-fitting it per K would return the same cut at K times the
-    price.  The gap #3116 identified is the missing arm above, not the hoist.
-
-    **The cost columns.**  ``fold_seconds`` is what #2897 and #3116 read: the
-    measured fit time of this K's own folds plus the count-independent overhead
-    of the threshold rule.  It is *not* the whole price of K, and #3314 found
-    that reading it as one under-states the exchange rate badly, because two
-    other K-proportional pieces of the shipped calibration are paid outside it
-    and outside every other timing column the frame carries:
-
-    * ``fold_score_seconds`` - one scoring pass over the sim set per fold, so
-      the shipped rule can anchor each fold's mixture on that fold's own
-      haystack (measured in :func:`_safe_threshold_for_step`).
-    * ``anchored_seconds`` - production's ``fit_fold_anchored_cut`` over the
-      prefix: one anchored EM per fold.
-
-    ``cal_seconds`` sums all of it (``fold_fit_seconds + overhead +
-    fold_score_seconds + anchored_seconds``) and is the column an affordability
-    ceiling has to read.  ``train_seconds``, ``xcal_seconds``,
-    ``pool_score_seconds`` and ``test_score_seconds`` cover the *rest* of the
-    step and none of them covers the safe-threshold block, so
-    ``cal_seconds + train_seconds + pool_score_seconds + test_score_seconds`` is
-    the per-step wall clock a live run at K would have.
-
-    All of it is measured inside the one Kmax run, so every K's timing shares
-    one machine, one process and one cache state - the *ratios* are the
-    load-bearing part, not the absolute seconds.
-
-    This is the study's screen, not its verdict, for the usual reason (see
-    :func:`_schedule_variant_rows`): K also steers acquisition through the
-    threshold Autopilot's Hard pick ranks around, and a screen that holds the
-    trajectory fixed cannot see the votes a different K would have collected.
-    The live A/B runs exist to measure exactly that.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.training.thresholds import (  # noqa: PLC0415
-        fit_gmm_threshold,
-        threshold_from_fold_orderings,
-    )
-
-    fold_data = details.get("fold_count_data")
-    if not fold_data:
-        return []
-    orderings = fold_data["orderings"]
-    seconds = fold_data["seconds"]
-    overhead = float(fold_data.get("overhead_seconds") or 0.0)
-    haystacks = fold_data.get("haystacks") or []
-    haystack_seconds = fold_data.get("haystack_seconds") or []
-    if not orderings:
-        return []
-
-    ctx = None
-    gmm_cut = gmm_fit = None
-    if sim_pooled_scores:
-        ctx = BlendContext(
-            n_labels=int(details["n_votes"]),
-            n_good=int(details["n_good"]),
-            n_bad=int(details["n_bad"]),
-        )
-        gmm_cut, gmm_fit = fit_gmm_threshold(sim_pooled_scores)
-
-    rows: list[dict[str, Any]] = []
-    for k in counts:
-        if k < 1 or k > len(orderings):
-            continue
-        prefix = orderings[:k]
-        cal_scores = np.array([s for scores, _ in prefix for s in scores])
-        cal_labels = np.array([lb for _, labels_ in prefix for lb in labels_])
-        xcal = threshold_from_fold_orderings(prefix, inclusion)
-        fold_fit_seconds = float(sum(seconds[:k]))
-        fold_seconds = _r(fold_fit_seconds + overhead)
-        # The rest of what a live run at K pays for, beside the fold fits
-        # (#3314).  Scoring the sim set once per fold is K-proportional and is
-        # measured in `_safe_threshold_for_step`; the anchored fit over the
-        # prefix is measured below, inside `_fold_count_arms`.
-        fold_score_seconds = float(sum(haystack_seconds[:k])) if len(haystack_seconds) >= k else float("nan")
-        timings: dict[str, float] = {}
-
-        arms = _fold_count_arms(
-            prefix,
-            xcal,
-            inclusion,
-            haystacks[:k] if len(haystacks) >= k else None,
-            sim_fit_scores if sim_fit_scores is not None else sim_pooled_scores,
-            ctx,
-            gmm_cut,
-            gmm_fit,
-            schedule,
-            timings=timings,
-        )
-        anchored_seconds = float(timings.get("anchored_seconds", float("nan")))
-        # The FULL calibration wall clock at K: fold fits + one haystack scoring
-        # pass per fold + production's anchored fit over the prefix + the
-        # count-independent overhead of the conformal rule.  `fold_seconds` is
-        # deliberately left as it was (#2897/#3116 read it, and archived runs
-        # are compared against it), but it is only the first of those four
-        # terms, so an affordability rule must read THIS column.
-        cal_seconds = _r(fold_fit_seconds + overhead + fold_score_seconds + anchored_seconds)
-
-        for arm, threshold, provenance, weight in arms:
-            row = _operating_metrics(
-                base_scores,
-                base_labels,
-                threshold,
-                inclusion,
-                cal_scores,
-                cal_labels,
-                pool_variant="max",
-                provenance=provenance,
-                n_pool_rows=n_pool_rows,
-            )
-            row["gmm_variant"] = f"folds_k{k}_{arm}"
-            row["schedule"] = schedule or ""
-            row["xcal_threshold"] = _r(xcal)
-            row["gmm_cut"] = _r(gmm_cut) if gmm_cut is not None else float("nan")
-            row["blend_weight"] = _r(weight)
-            row["fold_count"] = k
-            row["fold_seconds"] = fold_seconds
-            row["fold_fit_seconds"] = _r(fold_fit_seconds)
-            row["fold_score_seconds"] = _r(fold_score_seconds)
-            row["anchored_seconds"] = _r(anchored_seconds)
-            row["cal_seconds"] = cal_seconds
-            row["n_cal_scores"] = int(cal_scores.size)
-            row["n_folds_used"] = _folds_used(provenance, k)
-            rows.append(row)
-    return rows
-
-
-#: Default sweep grid for the anchored-mixture eval arms (issue #2852).  Each
-#: (anchor_weight, cut rule[, fold combine]) combination is one paired
-#: within-step variant; the GRID run overrides these via
-#: ``simulate_voting_iterations``'s ``anchored_*`` parameters to exhaust the
-#: grid registered in ``docs/plans/population-anchored-calibration.md``.
-_ANCHORED_WEIGHTS: tuple[float, ...] = (1.0, 10.0, 100.0)
-_ANCHORED_RULES: tuple[str, ...] = ("mid", "rate")
-_ANCHORED_FOLD_COMBINES: tuple[str, ...] = ("qmean",)
-
-
-def _anchored_variant_rows(
-    details: dict[str, Any],
-    base_scores: "np.ndarray",
-    base_labels: "np.ndarray",
-    sim_scores: list[float],
-    sim_ids: list[int],
-    good_ids: list[int],
-    bad_ids: list[int],
-    fold_haystacks: list,
-    inclusion: int,
-    n_pool_rows: float,
-    weights: list[float],
-    rules: list[str],
-    fold_combines: list[str],
-    fold_anchored: bool,
-    sim_fit_scores: list[float] | None = None,
-) -> list[dict[str, Any]]:
-    """Metric rows for the anchored-mixture threshold arms (issue #2852).
-
-    Three families, all evaluated against the same held-out test scores as the
-    #2799 variants so every row is step-paired with the shipped blend
-    (``pooled_mid``) and pure x-cal (``xcal_only``):
-
-    * ``anchored_w{W}_{rule}`` - the **label-anchored** mixture: anchored EM on
-      the final model's sim-set (haystack) scores with the voted items' own
-      final-model scores clamped to their labelled component.  One EM per
-      anchor weight; each cut rule re-cuts the same fit.
-    * ``fold_anchored_w{W}_{rule}_{combine}`` - the **fold-anchored**
-      ("cross-LabeledGMM") repair: per calibration fold, anchored EM on that
-      fold model's haystack scores with that fold's held-out labelled scores
-      as anchors (honest anchors, one shared scale), each fold's cut carried
-      to the final model by rank transfer and combined in quantile space.
-    * ``rank_transfer`` - the conformal x-cal cut carried from the pooled fold
-      haystack distribution to the final model's as a quantile: the
-      scale-transfer-only arm that attributes H1 (see the plan).
-
-    Anchored thresholds are used **raw** - the estimator replaces the blend
-    rather than feeding it - so ``blend_weight`` is NaN and ``raw_cut_*``
-    equals the headline cost columns.  The estimator path actually taken
-    (anchored / unanchored fallback / fold tally) is recorded in
-    ``threshold_provenance``.
-
-    *fold_haystacks* are the per-fold sim-set score arrays the shipped
-    threshold already computed (:func:`_safe_threshold_for_step`); the fold
-    family re-cuts them rather than paying the scoring passes twice.  The grid
-    point at the production ``(κ, rule, combine)`` therefore reproduces the
-    step's own shipped cut exactly - the grid's *other* points are the
-    deviation under test.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.calibration_metrics import inclusion_weights, operating_cost  # noqa: PLC0415
-
-    xcal = float(details["xcal_threshold"])
-    fold_orderings = details.get("fold_orderings") or []
-    cal_scores = np.array([s for scores, _ in fold_orderings for s in scores]) if fold_orderings else None
-    cal_labels = np.array([lb for _, labels_ in fold_orderings for lb in labels_]) if fold_orderings else None
-    wf, wn = inclusion_weights(inclusion)
-    nan = float("nan")
-
-    score_by_id = dict(zip(sim_ids, sim_scores, strict=True))
-    anchor_scores = [score_by_id[cid] for cid in (*good_ids, *bad_ids) if cid in score_by_id]
-    anchor_labels = [1.0] * sum(1 for cid in good_ids if cid in score_by_id) + [0.0] * sum(
-        1 for cid in bad_ids if cid in score_by_id
-    )
-    # The #3308 population convention: the voted items anchor the fits, so they
-    # are dropped from the free haystack sample instead of sitting in it twice
-    # (once free, once clamped) - matching the shipped cut and the (already
-    # filtered) *fold_haystacks* the fold family re-cuts.  *sim_fit_scores*
-    # carries the caller's floor decision (resolve_exclusion_floor), so this
-    # family can never diverge from the shipped population convention.
-    final_scores = np.asarray(sim_fit_scores if sim_fit_scores is not None else sim_scores, dtype=np.float64)
-
-    rows: list[dict[str, Any]] = []
-
-    def emit(name: str, threshold: float, provenance: str, cut_kind: str) -> None:
-        """Record one anchored arm.  *cut_kind* is the production rule's own
-        ``cut_fallback_kind`` (:func:`~vtscore.training.thresholds.gmm_cut_from_fit`),
-        so these rows say ``continued`` / ``degenerate_midpoint`` where the
-        decomposition family says ``midpoint`` - the two substitute different
-        values on the same fits (issue #2900).
-        """
-        if not np.isfinite(threshold):
-            return
-        row = _operating_metrics(
-            base_scores,
-            base_labels,
-            threshold,
-            inclusion,
-            cal_scores,
-            cal_labels,
-            pool_variant="max",
-            provenance=provenance,
-            n_pool_rows=n_pool_rows,
-        )
-        row["gmm_variant"] = name
-        row["xcal_threshold"] = _r(xcal)
-        row["gmm_cut"] = _r(threshold)
-        row["blend_weight"] = nan
-        row["cut_fallback"] = int(bool(cut_kind))
-        row["cut_fallback_kind"] = cut_kind
-        raw_cost, raw_fpr, raw_fnr = operating_cost(base_scores, base_labels, threshold, wf, wn)
-        row["raw_cut_cost"] = _r(raw_cost)
-        row["raw_cut_fpr"] = _r(raw_fpr)
-        row["raw_cut_fnr"] = _r(raw_fnr)
-        rows.append(row)
-
-    # --- Label-anchored family: one anchored EM per weight, re-cut per rule. ---
-    for weight in weights:
-        fit, provenance = anchored_gmm_fit(final_scores, anchor_scores, anchor_labels, anchor_weight=weight)
-        if fit is None:
-            continue
-        for rule in rules:
-            if rule in ("mid_tilt", "q_tilt"):
-                # Fold-level rules, defined in fold-quantile space: a single
-                # label-anchored fit has no folds to tilt across.  The fold
-                # family below sweeps them; here they are skipped rather than
-                # fed to gmm_cut_from_fit, which (correctly) rejects them.
-                continue
-            cut, cut_kind = gmm_cut_from_fit(fit, rule, wf, wn)
-            emit(f"anchored_w{weight:g}_{rule}", cut, provenance, cut_kind)
-
-    # --- Fold-anchored family + the rank-transfer attribution arm. ---
-    if fold_anchored and fold_haystacks and fold_orderings:
-        _emit_fold_anchored_rows(
-            emit,
-            xcal,
-            fold_haystacks,
-            fold_orderings,
-            final_scores,
-            inclusion,
-            weights,
-            rules,
-            fold_combines,
-        )
-
-    return rows
-
-
-def _emit_fold_anchored_rows(
-    emit: Callable[[str, float, str, str], None],
-    xcal: float,
-    fold_haystacks: list,
-    fold_orderings: list[tuple[list[float], list[float]]],
-    final_scores: "np.ndarray",
-    inclusion: int,
-    weights: list[float],
-    rules: list[str],
-    fold_combines: list[str],
-) -> None:
-    """Emit the fold-family arm rows over pre-computed per-fold sim scores.
-
-    The scoring pass per fold model is the fold arms' whole marginal cost, and
-    the shipped threshold already paid it (see :func:`_safe_threshold_for_step`),
-    so the grid re-cuts those arrays rather than re-scoring per grid point.
-    ``rank_transfer`` reuses them too: the conformal cut carried from the pooled
-    fold haystack distribution to the final model's - the scale-transfer-only
-    attribution arm of the plan.
-
-    Every grid point goes through the same
-    :func:`~vtscore.training.thresholds.fold_anchored_gmm_threshold` the app
-    ships; the grid *is* the deviation under test, so the arm at the production
-    (κ, rule, combine) reproduces the shipped cut exactly.
-
-    These rows carry no ``cut_fallback_kind``: a fold-anchored threshold is
-    composed from one per-fit cut *per fold* in quantile space, so there is no
-    single fit whose fallback branch the row could name.  A per-fold breakdown
-    would be a different column with a different unit of observation, and the
-    ``mid_tilt`` rule the app ships already degrades a rate-less fold to plain
-    ``mid`` rather than to a substituted value (see
-    :meth:`~vtscore.training.thresholds.FoldAnchoredCut._quantile_at`).
-    """
-    import numpy as np  # noqa: PLC0415
-
-    n_folds = min(len(fold_haystacks), len(fold_orderings))
-    fold_hay, orderings = fold_haystacks[:n_folds], fold_orderings[:n_folds]
-
-    emit(
-        "rank_transfer",
-        rank_transfer(xcal, np.concatenate(fold_hay), final_scores),
-        "rank_transfer",
-        CUT_KIND_INTERIOR,
-    )
-    for weight in weights:
-        for rule in rules:
-            for combine in fold_combines:
-                threshold, provenance = fold_anchored_gmm_threshold(
-                    fold_hay,
-                    orderings,
-                    final_scores,
-                    inclusion,
-                    anchor_weight=weight,
-                    cut_rule=rule,
-                    combine=combine,
-                )
-                emit(f"fold_anchored_w{weight:g}_{rule}_{combine}", threshold, provenance, CUT_KIND_INTERIOR)
 
 
 def _calibration_metric_rows(
@@ -1856,7 +566,7 @@ def _calibration_metric_rows(
     maybe_dump_predictions(clips_dict, list(ids), base_scores, list(labels), threshold, target_category)
     base_cal_scores = np.array([s for scores, _ in fold_orderings for s in scores]) if fold_orderings else None
     base_cal_labels = np.array([lb for _, labels_ in fold_orderings for lb in labels_]) if fold_orderings else None
-    base = _operating_metrics(
+    base = operating_metrics(
         base_scores,
         labels,
         threshold,
@@ -1870,7 +580,7 @@ def _calibration_metric_rows(
     if "xcal_threshold" in details:
         # Under safe_thresholds the base row's threshold is the blended one;
         # record the pre-blend conformal cut alongside it (issue #2799).
-        base["xcal_threshold"] = _r(float(details["xcal_threshold"]))
+        base["xcal_threshold"] = round6(float(details["xcal_threshold"]))
     # How many held-out scores the conformal quantile was actually taken over,
     # on the SHIPPED row rather than only on the fold-count variant rows (issue
     # #3287).  It was declared in `CALIBRATION_COLUMNS` and filled only by the
@@ -1914,7 +624,7 @@ def _calibration_metric_rows(
             v_cal_scores = np.array([s for scores, _ in v_orderings for s in scores])
             v_cal_labels = np.array([lb for _, labels_ in v_orderings for lb in labels_])
             rows.append(
-                _operating_metrics(
+                operating_metrics(
                     v_scores,
                     labels,
                     v_threshold,
@@ -1928,199 +638,6 @@ def _calibration_metric_rows(
             )
 
     return rows, base_scores, labels
-
-
-def _inclusion_sweep_rows(
-    details: dict[str, Any],
-    base_scores: "np.ndarray",
-    base_labels: "np.ndarray",
-    inclusion_sweep_ks: list[int],
-) -> list[dict[str, Any]]:
-    """Re-threshold the base fold orderings at each inclusion *k* and measure test FNR.
-
-    Near-free (no refits): pools the cached fold orderings once and applies the
-    conformal rule at each ``k``, then measures the realised test FPR/FNR at that
-    cut.  Checks the Inclusion budget ``alpha(k) = 0.25 * 2^-k`` against the
-    measured FNR under the **grouped** calibration path (issue #2781 / the
-    grouped-arm follow-up in inclusion-calibration-bias.md).  Returns ``[]`` when
-    the base threshold was a fallback (no real orderings to sweep).
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.calibration_metrics import inclusion_weights, operating_cost  # noqa: PLC0415
-
-    fold_orderings = details.get("fold_orderings") or []
-    if not fold_orderings:
-        return []
-    base_scores = np.asarray(base_scores, dtype=np.float64)
-    base_labels = np.asarray(base_labels, dtype=np.float64)
-    out: list[dict[str, Any]] = []
-    for k in inclusion_sweep_ks:
-        thr_k = threshold_from_fold_orderings(fold_orderings, k)
-        wf, wn = inclusion_weights(k)
-        cost_k, fpr_k, fnr_k = operating_cost(base_scores, base_labels, thr_k, wf, wn)
-        alpha_k = 0.25 * 2.0 ** (-k)
-        out.append(
-            {
-                "inclusion_k": k,
-                "alpha": _r(alpha_k),
-                "sweep_threshold": _r(float(thr_k)),
-                "sweep_fpr": _r(fpr_k),
-                "sweep_fnr": _r(fnr_k),
-                "excess_fnr": _r(fnr_k - alpha_k),
-            }
-        )
-    return out
-
-
-def _cut_inclusion_rows(
-    details: dict[str, Any],
-    base_scores: "np.ndarray",
-    base_labels: "np.ndarray",
-    fold_haystacks: list,
-    sim_scores: list[float],
-    ks: list[int],
-    weights: list[float],
-    rules: list[str],
-    fold_combines: list[str],
-    qtilt_steps: list[float],
-) -> list[dict[str, Any]]:
-    """Sweep the fold-anchored cut *rules* across the whole Inclusion knob (#2865).
-
-    The shipped ``mid`` cut was picked by two calibration runs that scored every
-    arm at inclusion 0, and a bare midpoint ignores the cost weights inclusion
-    arrives as - so it made the knob a no-op for every detector with usable
-    folds.  ``mid_tilt`` was shipped to restore the tilt while reproducing the
-    measured arm exactly at 0, but the *tilt* has never been priced against its
-    alternatives.  This frame is that measurement.
-
-    **Nearly free, and faithful for the same reason.**  The expensive part of a
-    fold-anchored threshold is the per-fold anchored EM, and it does not depend
-    on the cut rule, the combine, or the inclusion - so the fit is taken **once
-    per anchor weight** via the app's own :func:`fit_fold_anchored_cut`, and
-    every (rule, combine, k) point re-cuts it through the app's own
-    :meth:`~vtscore.training.thresholds.FoldAnchoredCut.threshold_at`.  That is
-    not merely cheaper than calling
-    :func:`~vtscore.training.thresholds.fold_anchored_gmm_threshold` per point,
-    it is exactly what production does when the user drags the slider
-    (:func:`vtscore.state.core.recompute_detector_thresholds_for_inclusion`
-    re-cuts a cached estimator with no refit), so the sweep measures the object
-    the app actually re-cuts rather than a chain of independent retrains.
-
-    A weight whose fit fails outright contributes no rows.  The terminal
-    fallbacks :func:`fold_anchored_gmm_threshold` applies in that case
-    (final-model unanchored midpoint, then its median) are deliberately *not*
-    reproduced here: both are inclusion-blind by construction, so they would
-    enter this frame as arms that trivially lose the knob-liveness comparison
-    while saying nothing about the rule under test.
-
-    Every row is scored under the cost weights of **its own** ``k`` and against
-    the oracle cut at that same ``k`` - the run's reporting inclusion does not
-    enter - so regret is comparable along the knob as well as across arms.
-
-    *sim_scores* arrives under the #3308 population convention (the caller
-    passes the voted-items-dropped haystack when it has one), matching the
-    already-filtered *fold_haystacks*, so every re-cut here realizes its
-    quantile on the same population the shipped cut does.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.calibration_metrics import inclusion_weights, operating_cost, oracle_cut  # noqa: PLC0415
-
-    fold_orderings = details.get("fold_orderings") or []
-    if not fold_orderings or not fold_haystacks or not ks:
-        return []
-
-    n_folds = min(len(fold_haystacks), len(fold_orderings))
-    fold_hay, orderings = fold_haystacks[:n_folds], fold_orderings[:n_folds]
-    final_scores = np.asarray(sim_scores, dtype=np.float64)
-    base_scores = np.asarray(base_scores, dtype=np.float64)
-    base_labels = np.asarray(base_labels, dtype=np.float64)
-    n_test = int(base_scores.size)
-
-    # The oracle depends only on k, not on the arm - hoist it out of the arm
-    # loop so a 5-rule x 8-weight grid pays for it once per k rather than 40x.
-    oracle_by_k = {}
-    for k in ks:
-        wf, wn = inclusion_weights(k)
-        o_thr, o_cost, _o_fpr, _o_fnr = oracle_cut(base_scores, base_labels, wf, wn)
-        oracle_by_k[k] = (wf, wn, o_thr, o_cost)
-
-    arms = _cut_inclusion_arms(fold_hay, orderings, final_scores, weights, rules, fold_combines, qtilt_steps)
-
-    out: list[dict[str, Any]] = []
-    for arm, arm_cut in arms:
-        for k in ks:
-            wf, wn, o_thr, o_cost = oracle_by_k[k]
-            thr = arm_cut.threshold_at(k)
-            if not np.isfinite(thr):
-                continue
-            cost, fpr, fnr = operating_cost(base_scores, base_labels, thr, wf, wn)
-            n_admitted = int(np.count_nonzero(base_scores >= thr))
-            out.append(
-                {
-                    **arm,
-                    "inclusion_k": k,
-                    "fold_quantile": _r(arm_cut.quantile_at(k)),
-                    "cut_threshold": _r(float(thr)),
-                    "cut_cost": _r(cost),
-                    "cut_fpr": _r(fpr),
-                    "cut_fnr": _r(fnr),
-                    "k_oracle_threshold": _r(float(o_thr)),
-                    "k_oracle_cost": _r(o_cost),
-                    "cut_regret": _r(cost - o_cost),
-                    "admitted_frac": _r(n_admitted / n_test if n_test else 0.0),
-                    "n_admitted": n_admitted,
-                    "n_test": n_test,
-                }
-            )
-    return out
-
-
-def _cut_inclusion_arms(
-    fold_hay: list,
-    orderings: list[tuple[list[float], list[float]]],
-    final_scores: "np.ndarray",
-    weights: list[float],
-    rules: list[str],
-    fold_combines: list[str],
-    qtilt_steps: list[float],
-) -> "list[tuple[dict[str, Any], Any]]":
-    """The ``(identity columns, re-cut estimator)`` pairs the #2865 sweep scores.
-
-    One anchored fit per *weight* - the only expensive step, and the only one
-    that depends on none of the swept axes - re-cut into every (rule, combine,
-    step) arm by :func:`dataclasses.replace`.  A weight whose fit fails
-    contributes no arms.
-    """
-    from vtscore.training.thresholds import fit_fold_anchored_cut  # noqa: PLC0415
-
-    arms: list[tuple[dict[str, Any], Any]] = []
-    for weight in weights:
-        cut = fit_fold_anchored_cut(fold_hay, orderings, final_scores, anchor_weight=weight)
-        if cut is None:
-            continue
-        for rule in rules:
-            # ``q_tilt`` alone carries a free step size, so it is the only rule
-            # that expands over ``qtilt_steps``; the others take a NaN step so
-            # the whole frame keeps one row shape.
-            steps = list(qtilt_steps) if rule == "q_tilt" else [float("nan")]
-            for combine in fold_combines:
-                for step in steps:
-                    arm_cut = replace(cut, cut_rule=rule, combine=combine)
-                    name = f"fold_anchored_w{weight:g}_{rule}_{combine}"
-                    if rule == "q_tilt":
-                        arm_cut = replace(arm_cut, qtilt_step=step)
-                        name += f"_s{step:g}"
-                    ident = {
-                        "arm": name,
-                        "cut_rule": rule,
-                        "anchor_weight": weight,
-                        "combine": combine,
-                        "qtilt_step": _r(step),
-                    }
-                    arms.append((ident, arm_cut))
-    return arms
 
 
 # ------------------------------------------------------------------
@@ -2251,7 +768,7 @@ def _skyline_arm_rows(
         o_thr, _o_cost, _o_fpr, _o_fnr = cm.oracle_cut(scores, test_labels, wf, wn)
         if not np.isfinite(o_thr):
             return None
-        row = _operating_metrics(
+        row = operating_metrics(
             scores,
             test_labels,
             float(o_thr),
@@ -2428,8 +945,8 @@ def _apply_skyline_decomposition(rows: list[dict[str, Any]], skyline_rows: list[
     for row in [*rows, *skyline_rows]:
         row["skyline_oracle_cost"] = floor
         row["skyline_oracle_cost_honest"] = floor_honest
-        row["training_regret"] = _r(row["oracle_cost"] - floor)
-        row["training_regret_honest"] = _r(row["oracle_cost_honest"] - floor_honest)
+        row["training_regret"] = round6(row["oracle_cost"] - floor)
+        row["training_regret_honest"] = round6(row["oracle_cost_honest"] - floor_honest)
 
 
 # ------------------------------------------------------------------
@@ -3191,7 +1708,7 @@ def simulate_voting_iterations(  # noqa: C901
                     "startup_round": startup_round,
                     "startup_held": bool(startup_state.held_for_quorum) if startup_state is not None else False,
                     "startup_extended_clicks": int(startup_state.extended_clicks) if startup_state is not None else 0,
-                    "startup_cut": _r(startup_cut) if startup_cut is not None else float("nan"),
+                    "startup_cut": round6(startup_cut) if startup_cut is not None else float("nan"),
                     "startup_cut_percentile": (
                         _sorted_percentile(seed_sorted_scores, startup_cut) if startup_cut is not None else float("nan")
                     ),
@@ -3199,11 +1716,13 @@ def simulate_voting_iterations(  # noqa: C901
                     "picked_label": 1 if is_positive else 0,
                     "picked_seed_rank": rank,
                     "picked_seed_percentile": (
-                        _r(rank / (n_sorted - 1)) if n_sorted > 1 and rank >= 0 else float("nan")
+                        round6(rank / (n_sorted - 1)) if n_sorted > 1 and rank >= 0 else float("nan")
                     ),
-                    "picked_seed_score": _r(seed_scores[cid]) if seed_scores and cid in seed_scores else float("nan"),
-                    "picked_detector_score": _r(pool_scores[cid]) if cid in pool_scores else float("nan"),
-                    "acq_threshold": _r(acq_threshold),
+                    "picked_seed_score": round6(seed_scores[cid])
+                    if seed_scores and cid in seed_scores
+                    else float("nan"),
+                    "picked_detector_score": round6(pool_scores[cid]) if cid in pool_scores else float("nan"),
+                    "acq_threshold": round6(acq_threshold),
                     "n_good": len(good_votes),
                     "n_bad": len(bad_votes),
                     "n_pool": len(pool),
